backend/requirement.txt
fastapi
"uvicorn[standard]"
sqlalchemy
pydantic-settings
psycopg2